MineNote : modify damage (int -> double)

+++ src/bms/model/MineNote.java
@@ -10,9 +10,9 @@ public class MineNote extends Note {
 	/**
 	 * 地雷のダメージ量
 	 */
-	private int damage;
+	private double damage;
 	
-	public MineNote(int wav, int damage) {
+	public MineNote(int wav, double damage) {
 		this.setWav(wav);
 		this.setDamage(damage);
 	}
@@ -21,7 +21,7 @@ public class MineNote extends Note {
 	 * 地雷ノーツのダメージ量を取得する
 	 * @return 地雷ノーツのダメージ量
 	 */
-	public int getDamage() {
+	public double getDamage() {
 		return damage;
 	}
 
@@ -29,7 +29,7 @@ public class MineNote extends Note {
 	 * 地雷ノーツのダメージ量を設定する
 	 * @param damage 地雷ノーツのダメージ量
 	 */
-	public void setDamage(int damage) {
+	public void setDamage(double damage) {
 		this.damage = damage;
 	}
